Tested stronger bricks

+++ BrickBreaker/components/ballComponent.hpp
@@ -11,7 +11,7 @@ namespace components
 	class Ball : public PolymorphicComparable<Component, Ball>
 	{
 	public:
-		Ball(float speed, glm::vec2 direction, float strength, entities::EntityPtr entity, bool isAttachedToPaddle = false) : 
+		Ball(float speed, glm::vec2 direction, int strength, entities::EntityPtr entity, bool isAttachedToPaddle = false) : 
 			speed(speed),
 			direction(direction),
 			strength(strength),
@@ -21,7 +21,7 @@ namespace components
 
 		float speed;
 		glm::vec2 direction;
-		float strength;
+		int strength;
 		bool isAttachedToPaddle; // This is for when you start a round and have to shoot the ball off
 		entities::EntityPtr paddle;
